dom/base/test/test_focus_scrollable_input.html

<!doctype html>
<title>Test for bug 1617342</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<style>
  input {
    /* Make them small enough that any value overflows */
    height: 10px;
    width: 10px;
    box-sizing: border-box;
    font: 16px/1 monospace;
  }
</style>
<input type="text" id="start" value="aaaaaaaaaa">

<input type="text" value="aaaaaaaaaa">
<input type="number" value="1111">
<input type="search" value="aaaaaaa">
<input type="url" value="https://fooooooooooo">

<input type="text" id="end" value="aaaaaaaaaa">
<script>
  SimpleTest.waitForExplicitFinish();
  SimpleTest.waitForFocus(async function() {
    // Enable Full Keyboard Access emulation on Mac.
    if (navigator.platform.indexOf("Mac") === 0) {
      await SpecialPowers.pushPrefEnv({"set": [["accessibility.tabfocus", 7]]});
    }

    const start = document.getElementById("start");

    start.focus();

    const end = document.getElementById("end");

    is(document.activeElement, start, "Focus moved sanely");

    let lastActiveElement = start;
    let stack = [start];

    do {
      synthesizeKey("KEY_Tab");
      isnot(document.activeElement, lastActiveElement, "Focus should've moved once per tab keypress");
      lastActiveElement = document.activeElement;
      stack.push(lastActiveElement);
    } while (document.activeElement != end)

    do {
      let previous = stack.pop();
      is(document.activeElement, previous, "Focus should've moved backwards as expected");
      synthesizeKey("KEY_Tab", {shiftKey: true});
    } while (stack.length);

    SimpleTest.finish();
  });
</script>
